+++ libnatools/natools.c
@@ -2,7 +2,7 @@
 
 #include "natools/natools.h"
 
-interval_t *interval_create(double x0, double x1) {
+interval_t *interval_create(long double x0, long double x1) {
   interval_t *i = (interval_t*)malloc(sizeof(interval_t));
   i->x0 = x0;
   i->x1 = x1;
@@ -15,12 +15,12 @@ void interval_destroy(interval_t *i) {
 }
 
 void interval_swap(interval_t *i) {
-  double m = i->x0;
+  long double m = i->x0;
   i->x0 = i->x1;
   i->x1 = m;
 }
 
-stop_cond_t *stop_cond_create(double e, int max_i) {
+stop_cond_t *stop_cond_create(long double e, int max_i) {
   stop_cond_t *s = (stop_cond_t*)malloc(sizeof(stop_cond_t));
   s->epsilon = e;
   s->max_iterations = max_i;
@@ -32,13 +32,13 @@ void stop_cond_destroy(stop_cond_t *s) {
     free(s);
 }
 
-double *vector_at(vector_t *v, size_t n) {
+long double *vector_at(vector_t *v, size_t n) {
   return v->a + n;
 }
 
 matrix_t *matrix_create(size_t r, size_t c) {
   matrix_t *m = (matrix_t*)malloc(sizeof(matrix_t));
-  m->a = (double*)malloc(sizeof(double) * r * c);
+  m->a = (long double*)malloc(sizeof(long double) * r * c);
   m->r = r;
   m->c = c;
   return m;
@@ -52,7 +52,7 @@ void matrix_destroy(matrix_t *m) {
   }
 }
 
-double *matrix_at(matrix_t *m, size_t r, size_t c) {
+long double *matrix_at(matrix_t *m, size_t r, size_t c) {
   return m->a + r * m->c + c;
 }
