+++ libnatools/natools/interpolation.h
@@ -4,9 +4,9 @@
 #include "natools/natools.h"
 #include "natools/function.h"
 
-int interpolate_newton(vector2_t *x, int n, double x0, double *r);
-int finite_difference(function_t *f, int n, double x0, double h, double *r);
-int interpolate_lagrange(vector2_t *x, int n, double x0, double *r);
+int interpolate_newton(vector2_t *x, int n, long double x0, long double *r);
+int finite_difference(function_t *f, int n, long double x0, long double h, long double *r);
+int interpolate_lagrange(vector2_t *x, int n, long double x0, long double *r);
 
 #endif /* _INTERPOLATION_H_ */
